+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Innmind\BlackBox\Application::filterOnTags()`
 - `Innmind\BlackBox\PHPUnit\Framework\Attributes\DataProvider`
 - `Innmind\BlackBox\PHPUnit\Framework\Attributes\Group`
+- `Innmind\BlackBox\Set::via()`
 
 ## 6.11.0 - 2026-03-18